moved all club pg images to public and renamed all paths
