+++ public/assests/clubs/logo/GDSC-IITI-Logo.png
