+++ src/app/club/clubHeads/page.tsx
@@ -4,31 +4,31 @@ import { motion, AnimatePresence } from "framer-motion";
 import React, { useState } from "react";
 import ScrollReveal from "./ScrollReveal";
 import Image from "next/image";
-import heroImage from "./assets/hero.png";
+import heroImage from "../../../../public/assests/clubHeads/hero.png";
 // import clubArr from "./clubData";
 import Header from "./header";
-import testImg from "./assets/img.png";
-import insta_icon from "./assets/social_logo/ig_black.webp";
-import li_icon from "./assets/social_logo/li_blackLogo.png";
-import site_icon from "./assets/social_logo/site.webp";
+import testImg from "../../../../public/assests/clubHeads/img.png";
+import insta_icon from "../../../../public/assests/clubHeads/social_logo/ig_black.webp";
+import li_icon from "../../../../public/assests/clubHeads/social_logo/li_blackLogo.png";
+import site_icon from "../../../../public/assests/clubHeads/social_logo/site.webp";
 
-import cae from "./assets/photos/CAE.jpg";
-import cfa_fin from "./assets/photos/CFA(Finance Department).jpg";
-import cfa_overall from "./assets/photos/CFA(Overall and Analytics Head).jpg";
-import concreate from "./assets/photos/Concreate.jpg";
-// import electr from "./assets/photos/";
-import cyn from "./assets/photos/Cynaptics.jpg";
-import ivdc from "./assets/photos/IVDC.jpg";
-import gdsc from "./assets/photos/GDSC.jpeg";
-// import meta from "./assets/photos/";
-import pclub_cp from "./assets/photos/PClub(Competitive Programming).jpeg";
-import pclub_dev from "./assets/photos/PClub(software dev).jpg";
-import pclub_cs from "./assets/photos/PClub(Cybersecurity).jpeg";
-// import quantum from "./assets/photos/";
-import robo from "./assets/photos/Robotics.jpg";
-// import snt from "./assets/photos/";
-import astro from "./assets/photos/TheAstronomyClub.jpg";
-import aero from "./assets/photos/Aeromodeling Club_.jpg";
+import cae from "../../../../public/assests/clubHeads/photos/CAE.jpg";
+import cfa_fin from "../../../../public/assests/clubHeads/photos/CFA(Finance Department).jpg";
+import cfa_overall from "../../../../public/assests/clubHeads/photos/CFA(Overall and Analytics Head).jpg";
+import concreate from "../../../../public/assests/clubHeads/photos/Concreate.jpg";
+// import electr from "../../../../public/assests/clubHeads/photos/";
+import cyn from "../../../../public/assests/clubHeads/photos/Cynaptics.jpg";
+import ivdc from "../../../../public/assests/clubHeads/photos/IVDC.jpg";
+import gdsc from "../../../../public/assests/clubHeads/photos/GDSC.jpeg";
+// import meta from "../../../../public/assests/clubHeads/photos/";
+import pclub_cp from "../../../../public/assests/clubHeads/photos/PClub(Competitive Programming).jpeg";
+import pclub_dev from "../../../../public/assests/clubHeads/photos/PClub(software dev).jpg";
+import pclub_cs from "../../../../public/assests/clubHeads/photos/PClub(Cybersecurity).jpeg";
+// import quantum from "../../../../public/assests/clubHeads/photos/";
+import robo from "../../../../public/assests/clubHeads/photos/Robotics.jpg";
+// import snt from "../../../../public/assests/clubHeads/photos/";
+import astro from "../../../../public/assests/clubHeads/photos/TheAstronomyClub.jpg";
+import aero from "../../../../public/assests/clubHeads/photos/Aeromodeling Club_.jpg";
 import test, { before } from "node:test";
 
 export default function Hero() {
