+++ src/app/club/clubs/page.tsx
@@ -3,23 +3,23 @@ import { motion, AnimatePresence } from "framer-motion";
 import React, { useState } from "react";
 import ScrollReveal from "./ScrollReveal";
 import Image from "next/image";
-import heroImage from "./assets/hero-new.jpg";
+import heroImage from "../../../../public/assests/clubs/hero-new.jpg";
 // import clubArr from "./clubData";
 import Header from "./header";
-import cae from "./assets/logo/cae.png";
-import cfa from "./assets/logo/CFA Club Logo.jpg";
-import concreate from "./assets/logo/concreate logo (new).jpg";
-import electr from "./assets/logo/Copy of electronics club (18).png";
-import cyn from "./assets/logo/Cynaptics club 20230325_104746.jpg";
-import gdsc from "./assets/logo/GDSC-IITI-Logo.png";
-import ivdc from "./assets/logo/IVDC Club 20230426_231624.jpg";
-import meta from "./assets/logo/MetaCryst Logo1.png";
-import pclub from "./assets/logo/Pclub.png";
-import quantum from "./assets/logo/quantum computing.jpg";
-import robo from "./assets/logo/ROBOTICS CLUB_.jpg";
-import snt from "./assets/logo/Science and technology council 23-24 20230515_112204 (1) (1).jpg";
-import astro from "./assets/logo/The_Astronomy_Club.jpg";
-import webgym from "./assets/logo/Web_Team_Logo.png";
+import cae from "../../../../public/assests/clubs/logo/cae.png";
+import cfa from "../../../../public/assests/clubs/logo/CFA Club Logo.jpg";
+import concreate from "../../../../public/assests/clubs/logo/concreate logo (new).jpg";
+import electr from "../../../../public/assests/clubs/logo/Copy of electronics club (18).png";
+import cyn from "../../../../public/assests/clubs/logo/Cynaptics club 20230325_104746.jpg";
+import gdsc from "../../../../public/assests/clubs/logo/GDSC-IITI-Logo.png";
+import ivdc from "../../../../public/assests/clubs/logo/IVDC Club 20230426_231624.jpg";
+import meta from "../../../../public/assests/clubs/logo/MetaCryst Logo1.png";
+import pclub from "../../../../public/assests/clubs/logo/Pclub.png";
+import quantum from "../../../../public/assests/clubs/logo/quantum computing.jpg";
+import robo from "../../../../public/assests/clubs/logo/ROBOTICS CLUB_.jpg";
+import snt from "../../../../public/assests/clubs/logo/Science and technology council 23-24 20230515_112204 (1) (1).jpg";
+import astro from "../../../../public/assests/clubs/logo/The_Astronomy_Club.jpg";
+import webgym from "../../../../public/assests/clubs/logo/Web_Team_Logo.png";
 
 import HeadsPage from "../clubHeads/page";
